requirements.txt
astroid==2.4.2
autopep8==1.5.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
idna==2.10
install==1.3.4
ipykernel==5.4.3
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
jupyter-client==6.1.11
jupyter-core==4.7.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pylint==2.6.2
python-dateutil==2.8.1
pyzmq==22.0.3
requests==2.25.1
Rx==3.1.1
SQLAlchemy==1.3.23
toml==0.10.2
tornado==6.1
twitch-python==0.0.19
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
python-Levenshtein==0.12.1
